+++ components/sound/openal_device/sources/openal_function_sample_decoder.cpp
@@ -39,7 +39,7 @@ FunctionSampleDecoder::~FunctionSampleDecoder ()
    Чтение декодированного звука количеством samples_count сэмплов начиная с first sample в data
 */
 
-size_t FunctionSampleDecoder::Read (size_t first_sample, size_t samples_count, void* data)
+unsigned int FunctionSampleDecoder::Read (unsigned int first_sample, unsigned int samples_count, void* data)
 {
   try
   {
